requirements.txt
regex>=2023.0.0
PyYAML>=6.0
